tools/requirements.txt
littlefs-python==0.17.1
